requirements.txt
# 基础依赖
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Mail==0.9.1
Flask-Login==0.6.3
google-generativeai==0.3.2
Pillow==10.1.0
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==2.0.23
Werkzeug==3.0.1
gunicorn==21.2.0  # 用于生产环境部署
redis==5.0.1      # 用于缓存系统
python-magic==0.4.27  # 用于文件类型检测
gevent==23.9.1    # 用于异步处理
supervisor==4.2.5  # 用于进程管理
psutil==5.9.6     # 用于系统资源监控
matplotlib>=3.8.2  # 用于数据可视化

# 基础NLP支持
jieba>=0.42.1     # 中文分词
nltk>=3.8.1       # 英文分词
tinysegmenter>=0.4  # 日文分词
langdetect>=1.0.9  # 语言检测
spacy>=3.5.0      # 高级NLP支持
pythainlp>=3.1.1  # 泰语支持
konlpy>=0.6.0     # 韩语支持
morfessor>=2.0.6  # 形态分析

# 网页处理
bs4>=4.12.2       # 用于自动更新在线字典

# 日语支持
fugashi>=2.2.1    # 日语形态分析
unidic-lite>=1.0.8  # 日语词典数据
mecab-python3>=1.0.6  # 日语分词增强

# 韩语支持
JPype1>=1.4.1     # 用于韩语支持

# 中文增强
scikit-learn>=1.3.0  # 用于文本聚类和相似度计算

# 词向量和语义分析
gensim>=4.3.2     # 词向量支持
rank-bm25>=0.2.2  # 文本相关性排序

# 依赖包
regex>=2023.10.3  # 增强的正则表达式支持
pyyaml>=6.0.1

# 特征权重词典更新相关依赖
SPARQLWrapper>=2.0.0  # SPARQL查询支持
rdflib>=7.0.0     # RDF数据处理
networkx>=3.2.1   # 图数据处理
tqdm>=4.65.0      # 进度条显示
pandas>=2.0.3     # 数据处理
urllib3>=2.0.7    # HTTP客户端
requests-cache>=1.1.0  # 请求缓存
retry>=0.9.2      # 重试机制
aiohttp>=3.8.6    # 异步HTTP请求
